Move printing to one file
