+++ cmd/beaker/experiment.go
@@ -298,60 +298,3 @@ func canonicalizeSpecV1(spec *api.ExperimentSpecV1) error {
 	}
 	return nil
 }
-
-func printExperiments(experiments []api.Experiment) error {
-	switch format {
-	case formatJSON:
-		return printJSON(experiments)
-	default:
-		if err := printTableRow(
-			"ID",
-			"WORKSPACE",
-			"AUTHOR",
-			"CREATED",
-			"STATUS",
-			"ARCHIVED",
-		); err != nil {
-			return err
-		}
-		for _, experiment := range experiments {
-			name := experiment.ID
-			if experiment.Name != "" {
-				name = experiment.Name
-			}
-			var archived string
-			if experiment.Archived {
-				archived = "archived"
-			}
-			if err := printTableRow(
-				name,
-				experiment.Workspace.Name,
-				experiment.Author.Name,
-				experiment.Created,
-				experimentStatus(experiment),
-				archived,
-			); err != nil {
-				return err
-			}
-		}
-		return nil
-	}
-}
-
-func experimentStatus(experiment api.Experiment) string {
-	counts := make(map[string]int)
-	for _, execution := range experiment.Executions {
-		status := executionStatus(execution.State)
-		count, ok := counts[status]
-		if ok {
-			counts[status] = count + 1
-		} else {
-			counts[status] = 1
-		}
-	}
-	var parts []string
-	for status, count := range counts {
-		parts = append(parts, fmt.Sprintf("%d %s", count, status))
-	}
-	return strings.Join(parts, ", ")
-}
