+++ cmd/beaker/group.go
@@ -247,40 +247,3 @@ func trimAndUnique(ids []string) []string {
 
 	return unique
 }
-
-func printGroups(groups []api.Group) error {
-	switch format {
-	case formatJSON:
-		return printJSON(groups)
-	default:
-		if err := printTableRow(
-			"ID",
-			"WORKSPACE",
-			"AUTHOR",
-			"CREATED",
-			"ARCHIVED",
-		); err != nil {
-			return err
-		}
-		for _, group := range groups {
-			name := group.ID
-			if group.Name != "" {
-				name = group.Name
-			}
-			var archived string
-			if group.Archived {
-				archived = "archived"
-			}
-			if err := printTableRow(
-				name,
-				group.Workspace.Name,
-				group.Author.Name,
-				group.Created,
-				archived,
-			); err != nil {
-				return err
-			}
-		}
-		return nil
-	}
-}
